refactor(applet/notification): rename helpers.h to notification_common.h
Avoid potential conflicts.
