+++ src/applet/notification/notification_common.c
@@ -1,4 +1,4 @@
-#include "helpers.h"
+#include "notification_common.h"
 
 #include <ctype.h>
 #include <string.h>
