+++ src/applet/notification/notification_common.h
