+++ src/applet/notification/replay.c
@@ -1,4 +1,4 @@
-#include "helpers.h"
+#include "notification_common.h"
 #include "process.h"
 
 #include <euicc/es10b.h>
